fix ci

+++ setup.py
@@ -103,9 +103,10 @@ class CMakeBuild(build_ext):
         # cp -rf $GITHUB_WORKSPACE/Release-build/lib SVF-${osVersion}/Release-build/
         # cp -rf $GITHUB_WORKSPACE/Release-build/bin SVF-${osVersion}/Release-build/
         shutil.copytree(os.path.join(SVF_DIR, CMAKE_BUILD_TYPE+"-build", "include"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "include"),dirs_exist_ok=True)
-        shutil.copytree(os.path.join(SVF_DIR, "svf", "include"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "include"),dirs_exist_ok=True)
-
-        shutil.copytree(os.path.join(SVF_DIR, "svf-llvm", "include"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "include"),dirs_exist_ok=True)
+        if os.path.exists(os.path.join(SVF_DIR, "svf", "include")):
+            shutil.copytree(os.path.join(SVF_DIR, "svf", "include"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "include"),dirs_exist_ok=True)
+        if os.path.exists(os.path.join(SVF_DIR, "svf-llvm", "include")):
+            shutil.copytree(os.path.join(SVF_DIR, "svf-llvm", "include"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "include"),dirs_exist_ok=True)
         shutil.copytree(os.path.join(SVF_DIR, CMAKE_BUILD_TYPE+"-build", "lib"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "lib"),dirs_exist_ok=True)
         shutil.copytree(os.path.join(SVF_DIR, CMAKE_BUILD_TYPE+"-build", "bin"), os.path.join(self.build_lib, "pysvf", "SVF", "Release-build", "bin"),dirs_exist_ok=True)
